Proofread
Refactor code, proofread and grammar check.

Remove one example. It more or less repeated the previous full example.
