+++ docs/csharp/programming-guide/delegates/how-to-declare-instantiate-and-use-a-delegate.md
@@ -1,76 +1,41 @@
 ---
 title: "How to declare, instantiate, and use a delegate"
-description: Learn how to declare, instantiate, and use a delegate. See examples that cover C# 1.0, 2.0, and 3.0 and later.
+description: Learn how to declare, instantiate, and use a delegate. This article provides several examples of declaring, instantiating, and invoking delegates.
 ms.topic: how-to
-ms.date: 07/20/2015
+ms.date: 12/20/2024
 helpviewer_keywords: 
   - "delegates [C#], declaring and instantiating"
-ms.assetid: 61c4895f-f785-48f8-8bfe-db73b411c4ae
 ---
 # How to declare, instantiate, and use a Delegate (C# Programming Guide)
 
 You can declare delegates using any of the following methods:
 
 - Declare a delegate type and declare a method with a matching signature:
-  
- [!code-csharp[csProgGuideDelegates#13](./snippets/Delegates.cs#13)]  
-  
- [!code-csharp[csProgGuideDelegates#14](./snippets/Delegates.cs#14)]  
-  
-- Assign a method group to a delegate type:
-  
- [!code-csharp[csProgGuideDelegates#32](./snippets/Delegates.cs#32)]  
-  
-- Declare an anonymous method:
-  
- [!code-csharp[csProgGuideDelegates#15](./snippets/Delegates.cs#15)]  
-  
-- Use a lambda expression:
-  
- [!code-csharp[csProgGuideDelegates#31](./snippets/Delegates.cs#31)]  
-  
- For more information, see [Lambda Expressions](../../language-reference/operators/lambda-expressions.md).  
-  
- The following example illustrates declaring, instantiating, and using a delegate. The `BookDB` class encapsulates a bookstore database that maintains a database of books. It exposes a method, `ProcessPaperbackBooks`, which finds all paperback books in the database and calls a delegate for each one. The `delegate` type that is used is named `ProcessBookCallback`. The `Test` class uses this class to print the titles and average price of the paperback books.  
-  
- The use of delegates promotes good separation of functionality between the bookstore database and the client code. The client code has no knowledge of how the books are stored or how the bookstore code finds paperback books. The bookstore code has no knowledge of what processing is performed on the paperback books after it finds them.  
-  
-## Example  
 
- [!code-csharp[csProgGuideDelegates#12](./snippets/Delegates.cs#12)]  
-  
-## Work with delegates
-  
-- Declaring a delegate.  
-  
-     The following statement declares a new delegate type.  
-  
-     [!code-csharp[csProgGuideDelegates#16](./snippets/Delegates.cs#16)]  
-  
-     Each delegate type describes the number and types of the arguments, and the type of the return value of methods that it can encapsulate. Whenever a new set of argument types or return value type is needed, a new delegate type must be declared.  
-  
-- Instantiating a delegate.  
-  
-     After a delegate type has been declared, a delegate object must be created and associated with a particular method. In the previous example, you do this by passing the `PrintTitle` method to the `ProcessPaperbackBooks` method as in the following example:  
-  
-     [!code-csharp[csProgGuideDelegates#17](./snippets/Delegates.cs#17)]  
-  
-     This creates a new delegate object associated with the [static](../../language-reference/keywords/static.md) method `Test.PrintTitle`. Similarly, the non-static method `AddBookToTotal` on the object `totaller` is passed as in the following example:  
-  
-     [!code-csharp[csProgGuideDelegates#18](./snippets/Delegates.cs#18)]  
-  
-     In both cases a new delegate object is passed to the `ProcessPaperbackBooks` method.  
-  
-     After a delegate is created, the method it is associated with never changes; delegate objects are immutable.  
-  
-- Calling a delegate.  
-  
-     After a delegate object is created, the delegate object is typically passed to other code that will call the delegate. A delegate object is called by using the name of the delegate object, followed by the parenthesized arguments to be passed to the delegate. Following is an example of a delegate call:  
-  
-     [!code-csharp[csProgGuideDelegates#19](./snippets/Delegates.cs#19)]  
-  
-     A delegate can be either called synchronously, as in this example, or asynchronously by using `BeginInvoke` and `EndInvoke` methods.  
-  
+  :::code language="csharp" source="./snippets/HowToDeclareAndUse.cs" id="DeclareNamedDelegate":::
+
+  :::code language="csharp" source="./snippets/HowToDeclareAndUse.cs" id="CreateNamedInstance":::
+
+- Assign a method group to a delegate type:
+
+  :::code language="csharp" source="./snippets/HowToDeclareAndUse.cs" id="MethodGroup":::
+
+- Declare an anonymous method
+
+  :::code language="csharp" source="./snippets/HowToDeclareAndUse.cs" id="AnonymousMethod":::
+
+- Use a lambda expression:
+
+  :::code language="csharp" source="./snippets/HowToDeclareAndUse.cs" id="LambdaExpression":::
+
+For more information, see [Lambda Expressions](../../language-reference/operators/lambda-expressions.md).
+
+The following example illustrates declaring, instantiating, and using a delegate. The `BookDB` class encapsulates a bookstore database that maintains a database of books. It exposes a method, `ProcessPaperbackBooks`, which finds all paperback books in the database and calls a delegate for each one. The `delegate` type is named `ProcessBookCallback`. The `Test` class uses this class to print the titles and average price of the paperback books.
+
+The use of delegates promotes good separation of functionality between the bookstore database and the client code. The client code has no knowledge of how the books are stored or how the bookstore code finds paperback books. The bookstore code has no knowledge of what processing is performed on the paperback books after it finds them.
+
+:::code language="csharp" source="./snippets/BookStore.cs":::
+
 ## See also
 
 - [Events](../events/index.md)
