+++ docs/csharp/programming-guide/delegates/index.md
@@ -1,11 +1,10 @@
 ---
 title: "Delegates"
 description: A delegate in C# is a type that refers to methods with a parameter list and return type. Delegates are used to pass methods as arguments to other methods.
-ms.date: 02/02/2021
+ms.date: 12/20/2024
 helpviewer_keywords: 
   - "C# language, delegates"
   - "delegates [C#]"
-ms.assetid: 97de039b-c76b-4b9c-a27d-8c1e1c8d93da
 ---
 # Delegates (C# Programming Guide)
 
@@ -13,7 +12,7 @@ A [delegate](../../language-reference/builtin-types/reference-types.md) is a typ
 
 Delegates are used to pass methods as arguments to other methods. Event handlers are nothing more than methods that are invoked through delegates. You create a custom method, and a class such as a windows control can call your method when a certain event occurs. The following example shows a delegate declaration:
 
-[!code-csharp[csProgGuideDelegates#20](./snippets/Delegates.cs#20)]
+:::code language="csharp" source="./snippets/Overview.cs" id="DelegateDeclaration":::
 
 Any method from any accessible class or struct that matches the delegate type can be assigned to the delegate. The method can be either static or an instance method. This flexibility means you can programmatically change method calls, or plug new code into existing classes.
