+++ docs/csharp/programming-guide/delegates/snippets/DelegateExamples.csproj
@@ -1,0 +1,11 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>true</ImplicitUsings>
+
+    <StartupObject>Bookstore.Test</StartupObject>
+  </PropertyGroup>
+</Project>
