+++ docs/csharp/programming-guide/delegates/snippets/Overview.cs
@@ -1,0 +1,8 @@
+﻿namespace DelegateExamples;
+
+public class OverViewExamples
+{
+    //<DelegateDeclaration>
+    public delegate int PerformCalculation(int x, int y);
+    //</DelegateDeclaration>
+}
